Revert "LPS-123758 Move classes that are not exported to internal package - portal"
This reverts commit df017eb5611f6287ba19470962d5ba88b0fd58f0.
