+++ modules/apps/portal/portal-upload/src/main/java/com/liferay/portal/upload/constants/LegacyUploadServletRequestPropsKeys.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.upload.internal.constants;
+package com.liferay.portal.upload.constants;
 
 /**
  * @author Pei-Jung Lan
